requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
Pillow
playwright
firecrawl-py
python-dotenv
requests
beautifulsoup4
aiofiles

